Fix some cli naming issues for clarity
