+++ src/madsci_common/madsci/common/backup_tools/__init__.py
@@ -3,7 +3,7 @@
 from .backup_manager import BackupManager
 from .backup_validator import BackupValidator
 from .base_backup import AbstractBackupTool, BackupInfo
-from .cli import main_mongodb_backup
+from .mongo_cli import main_mongodb_backup
 from .mongodb_backup import MongoDBBackupTool
 from .postgres_backup import PostgreSQLBackupTool
 from .postgres_cli import main_postgres_backup
